+++ packages/terra-form-checkbox/tests/wdio/form-checkbox-field-spec.js
@@ -38,7 +38,6 @@ Terra.describeViewports('CheckboxField', ['tiny'], () => {
 
   describe('CheckboxField Interactions - Invalid State', () => {
     before(() => {
-      browser.url('/#/raw/tests/terra-form-checkbox/form-checkbox/checkbox-field/controlled-checkbox-field');
       browser.click('[for="website-dept"]');
       browser.click('[for="ux-dept"]');
     });
@@ -59,7 +58,6 @@ Terra.describeViewports('CheckboxField', ['tiny'], () => {
 
   describe('CheckboxField Interactions - Invalid State with hideRequired', () => {
     before(() => {
-      browser.url('/#/raw/tests/terra-form-checkbox/form-checkbox/checkbox-field/hide-required-invalid-checkbox-field');
       browser.waitForVisible('#testing-checkbox-field');
       browser.click('[for="website-dept"]');
       browser.click('[for="ux-dept"]');